ISSUE-33: Refactor again
