+++ .gitignore
@@ -47,3 +47,6 @@ coverage.xml
 .venv/
 venv/
 ENV/
+
+# vscode
+settings.json
