+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn --worker-class eventlet -w 1 'web:app'
+web: gunicorn --worker-class eventlet -w 1 'web.server:app'
