+++ smart_canvas/__init__.py
@@ -1,3 +1,1 @@
 """ __init__.py """
-
-#from .core import CanvasCore
