+++ smart_canvas/__main__.py
@@ -5,7 +5,7 @@ import time
 import argparse
 
 # Internal packages
-from render import SmartRender
+from smart_canvas.render import SmartRender
 
 if __name__ == '__main__':
     SmartRender.run()
