+++ smart_canvas/filters/struc.pgm
