+++ smart_canvas/render.py
@@ -3,9 +3,9 @@ import time
 import moderngl
 import cv2
 
-from capture import VideoRead
-from core import CanvasCore
-from window import Window
+from smart_canvas.capture import VideoRead
+from smart_canvas.core import CanvasCore
+from smart_canvas.window import Window
 
 
 class SmartRender(Window):
@@ -21,10 +21,10 @@ class SmartRender(Window):
     def render(self, _time, frame_time):
         time.sleep(0.05)
         self.core.frame = self.video.frame
-        out_frame = self.core.frameout
+        out_frame = self.core.out_frame
         if (out_frame is None):
             return
-        self.frame_texture.write(cv2.flip(out_frame, 0))
+        self.frame_texture.write(cv2.flip(self.core.out_frame, 0))
         self.frame_texture.use(0)
         self.quad.render(mode=moderngl.TRIANGLE_STRIP)
